+++ R/plotsample.R
@@ -13,14 +13,15 @@
 # @export
 
 makepoly <- function(start, width, height) {
-  poly <- matrix(c(
-    start[1], start[2],
-    start[1], start[2] + height,
-    start[1] + width, start[2] + height,
-    start[1] + width, start[2],
-    start[1], start[2]
-  ),
-  ncol = 2, byrow = TRUE
+  poly <- matrix(
+    c(
+      start[1], start[2],
+      start[1], start[2] + height,
+      start[1] + width, start[2] + height,
+      start[1] + width, start[2],
+      start[1], start[2]
+    ),
+    ncol = 2, byrow = TRUE
   )
   return(Polygon(poly))
 }