+++ NEWS.md
@@ -4,6 +4,11 @@
 
 * Allow `NA` input for default 1D mappers to generate effect zero, like
   in `inla()`.
+  
+* New methods for `fm_crs`, `fm_CRS`, `fm_ellipsoid_radius`, `fm_transform`
+  to further support `sf` objects.
+  
+* Allow `sf` input to `family="cp"` models.
 
 * Further `bru_mapper` method updates;
 
@@ -24,6 +29,10 @@
   
   * Removed `methods` argument from `bru_mapper_define()`.  Implementations
     should register S3 methods instead.
+    
+  * Add `indexed` logical option to `bru_mapper_factor()`, to allow
+  factor inputs to be mapped to index values, as needed for `group` and
+  `replicate`. Fixes #174
 
 ## Bug fixes